pyproject.toml
[tool.poetry]
name = "papertrack"
version = "0.1.0"
description = ""
authors = ["joão danilo <joaodanilo1992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
peewee = "^3.14.8"
CherryPy = "^18.6.1"
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
